[FIX] Use unique accounts and improved bank search in base.
