+++ addons/account_bank_statement_import/__init__.py
@@ -1,4 +1,3 @@
 # -*- encoding: utf-8 -*-
 
-from . import res_partner_bank
 from . import account_bank_statement_import
