+++ openerp/addons/base/tests/test_res_partner_bank.py
